+++ src/backendAwsKeys.go
@@ -11,7 +11,7 @@ import (
 
 // get KeyPair
 func (d *backendAws) getKey(clusterName string) (keyName string, keyPath string, err error) {
-	keyName = fmt.Sprintf("aerolab-%s", clusterName)
+	keyName = fmt.Sprintf("aerolab-%s_%s", clusterName, a.opts.Config.Backend.Region)
 	keyPath = path.Join(string(a.opts.Config.Backend.SshKeyPath), keyName)
 	// check keyName exists, if not, error
 	filter := ec2.DescribeKeyPairsInput{}
@@ -33,7 +33,7 @@ func (d *backendAws) getKey(clusterName string) (keyName string, keyPath string,
 
 // get KeyPair
 func (d *backendAws) makeKey(clusterName string) (keyName string, keyPath string, err error) {
-	keyName = fmt.Sprintf("aerolab-%s", clusterName)
+	keyName = fmt.Sprintf("aerolab-%s_%s", clusterName, a.opts.Config.Backend.Region)
 	keyPath = path.Join(string(a.opts.Config.Backend.SshKeyPath), keyName)
 	_, _, err = d.getKey(clusterName)
 	if err == nil {
@@ -53,14 +53,14 @@ func (d *backendAws) makeKey(clusterName string) (keyName string, keyPath string
 		return
 	}
 	err = os.WriteFile(keyPath, []byte(*out.KeyMaterial), 0600)
-	keyName = fmt.Sprintf("aerolab-%s", clusterName)
+	keyName = fmt.Sprintf("aerolab-%s_%s", clusterName, a.opts.Config.Backend.Region)
 	keyPath = path.Join(string(a.opts.Config.Backend.SshKeyPath), keyName)
 	return
 }
 
 // get KeyPair
 func (d *backendAws) killKey(clusterName string) (keyName string, keyPath string, err error) {
-	keyName = fmt.Sprintf("aerolab-%s", clusterName)
+	keyName = fmt.Sprintf("aerolab-%s_%s", clusterName, a.opts.Config.Backend.Region)
 	keyPath = path.Join(string(a.opts.Config.Backend.SshKeyPath), keyName)
 	os.Remove(keyPath)
 	filter := ec2.DeleteKeyPairInput{}