+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 #### 4.2.0
 * New Features:
   * support arm deployments
-    * TODO: client create and grow, client base, client tools, aerospike upgrade, cluster create and grow (download, template creation, etc), installer download, template create, template delete, template list
-    * TODO: scripts.go - split and cleanup, this is a mess ; implement arm and amd flavours if needed
+    * TODO: aerospike upgrade, cluster create
     * TODO: insert/delete data (and build scripts)
-  * add `ams` client installation system
+    * TODO: add placeholder support for aerospike arm downloads to test (placeholder arm tgz output for all versions - dummy tgz)
+  * TODO: add `ams` client installation system
 * Improvements:
   * make mesh mode default
 * Bug fixes: